feat: add RecetaController and RecetaServiceImpl for recipe management
- Implemented RecetaController to handle HTTP requests related to recipes, including creating, updating, retrieving, and deleting recipes.
- Added RecetaServiceImpl to encapsulate business logic for recipe operations, including interaction with the RecetaRepository.
- Introduced new endpoint definitions and response handling for successful and error scenarios.
- Created a new TypeScript declaration file for Expo environment to ensure type safety in the frontend.

+++ backend/src/main/java/com/api/apos/domain/receta/entity/DetalleReceta.java
@@ -1,5 +1,7 @@
 package com.api.apos.domain.receta.entity;
 
+import java.math.BigDecimal;
+
 import com.api.apos.domain.material.Material;
 import com.api.apos.enums.Unidad;
 
@@ -21,12 +23,12 @@ public class DetalleReceta {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Float cantidad;
+    private BigDecimal cantidad;
 
     @Enumerated(EnumType.STRING)
     private Unidad unidadMedida;
 
-    private Float merma;
+    private BigDecimal merma;
 
     @ManyToOne
     @JoinColumn(name = "receta_id")
